test: integrate whitespace/bracket tests into existing test file and verify values
Integrated the whitespace and bracket parsing tests into `InetAddressParserTest.kt` instead of creating a new file to prevent test suite fragmentation. Updated the assertions to verify that the actual parsed values match cleanly parsed literals, ensuring the robustness logic correctly cleans inputs without mangling the data.
